+++ HISTORY.md
@@ -1,7 +1,8 @@
 # History
 
 # 0.8.0 (dev)
-
+- Fix styling for missing
+- Counts would break on completely null columns. Filter those out.
 # 0.7.0 (2020-8-21)
 - Add counts function to show the number of total and unique values in a column
 - Some doc cleanups and clarifications